A much-later version of GTest induces some true-but-not-serious signed-comparison warnings; eliminating the test code that exhibited these (correcting it).

+++ src/ipc/shm/arena_lend/test/divisible_shm_pool_test.cpp
@@ -57,7 +57,7 @@ TEST(Divisible_shm_pool_test, Interface)
     EXPECT_EQ(pool.get_id(), S_ID);
     EXPECT_EQ(pool.get_name(), S_NAME);
     EXPECT_EQ(pool.get_address(), S_ADDRESS);
-    EXPECT_EQ(pool.get_size(), S_SIZE);
+    EXPECT_EQ(size_t(pool.get_size()), S_SIZE);
     EXPECT_EQ(pool.get_fd(), S_FD);
     EXPECT_EQ(pool.get_remaining_size(), S_SIZE);
     EXPECT_FALSE(pool.is_completely_removed());
